Basic Lighting (WIP)
Very basic one-direction lighting (n dot L) lighting implemented in
shader
Vertex struct in GameManager now also takes an XMFLOAT3 for a normal
Crate movement now has delta time applied

+++ KVEngine/Source/Game/Spaceship.h
@@ -16,10 +16,10 @@ public:
 
 private:
 
-	float m_Vel;
-	float m_YawVel;
-	float m_PitchVel;
-	float m_RollVel;
+	double m_Vel;
+	double m_YawVel;
+	double m_PitchVel;
+	double m_RollVel;
 
 };
